requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
cycler==0.10.0
Django==2.1.2
djangorestframework==3.8.2
httplib2==0.11.3
kiwisolver==1.0.1
kombu==3.0.37
matplotlib==3.0.0
mysqlclient==1.3.12
numpy==1.15.1
pdfkit==0.6.1
pip-autoremove==0.9.1
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
vine==1.1.4
